+++ CHANGELOG.md
@@ -292,3 +292,4 @@
 - 2024-02-14: Log conf, leads, lib, bettercontact, +3
 - 2024-02-15: Shutdown lib, chat, descend, mod, +6
 - 2024-02-15: Integ README, campaign, qualify, chat, +2
+- 2024-02-16: Builder README, leads, contacts, task, +3